(REF) MixinTestCase - Move more code from template to base-class
